Modernize partner redux
