+++ src/redux/actions/index.ts
@@ -1,7 +1,6 @@
 import firebase from 'firebase';
 import { helperActions } from './helper';
 import { notificationsActions } from './notifications';
-import { partnersActions } from './partners';
 import { previousSpeakersActions } from './previous-speakers';
 import { routingActions } from './routing';
 import { scheduleActions } from './schedule';
@@ -12,7 +11,6 @@ import { userActions } from './user';
 export {
   helperActions,
   notificationsActions,
-  partnersActions,
   previousSpeakersActions,
   routingActions,
   scheduleActions,
