+++ src/redux/partners/state.ts
@@ -1,0 +1,9 @@
+import { PartnersState } from './types';
+
+export const initialPartnersState: PartnersState = {
+  fetching: false,
+  fetchingError: null,
+  list: [],
+  adding: false,
+  addingError: null,
+};
