+++ src/redux/reducer.ts
@@ -5,7 +5,7 @@ import { filtersReducer } from './filters/reducers';
 import { galleryReducer } from './gallery/reducers';
 import { initialState } from './initial-state';
 import { notificationsReducer } from './reducers/notifications';
-import { partnersReducer } from './reducers/partners';
+import { partnersReducer } from './partners/reducers';
 import { previousSpeakersReducer } from './reducers/previous-speakers';
 import { routingReducer } from './reducers/routing';
 import { scheduleReducer } from './reducers/schedule';
